+++ src/widget/file/filepicker.js
@@ -52,7 +52,7 @@ Filepicker.prototype._init = function() {
     this.$widget = $(
             '<div class="widget file-picker">' +
             '<input class="ignore fake-file-input"/>' +
-            '<button class="btn-icon-only btn-reset" type="button"><i class="icon icon-refresh"> </i></button>' +
+            ( this.props.readonly ? '' : '<button class="btn-icon-only btn-reset" type="button"><i class="icon icon-refresh"> </i></button>' ) +
             '<a class="btn-icon-only btn-download" download href=""><i class="icon icon-download"> </i></a>' +
             '<div class="file-feedback"></div>' +
             '<div class="file-preview"></div>' +